DDPG TF Policy: cast variable to tf.float32
This causes a type mismatch when running eagerly with
some environments. E.g.

    rllib train --run DDPG --env MountainCarContinuous-v0 --eager

+++ rllib/agents/ddpg/ddpg_tf_policy.py
@@ -227,7 +227,7 @@ def ddpg_actor_critic_loss(policy, model, _, train_batch):
         else:
             errors = 0.5 * tf.square(td_error)
 
-    critic_loss = tf.reduce_mean(train_batch[PRIO_WEIGHTS] * errors)
+    critic_loss = tf.reduce_mean(tf.cast(train_batch[PRIO_WEIGHTS], dtype=tf.float32) * errors)
     actor_loss = -tf.reduce_mean(q_t_det_policy)
 
     # Add l2-regularization if required.
